feat: add AbortController support to sendAndReceive (#54)
* feat: add AbortController support to sendAndReceive

The sendAndReceive functions now allow an AbortSignal to be passed
as an option. When passed, the default timeout is ignored. The caller
can then control whether to terminate the sendAndReceive call using a
timeout or some other reason.

Also fixes some documentation errors and some minor problems with the
demo.

Also adds github workflows