feat: propagate abort signal to message receiver
The initial AbortSignal implementation only cancelled
the chatty message sender. This change propagates abort
signals to the chatty message receiver. The caller MUST
opt in to propagate the signal as the receiving function
needs to be aware that the signal will be passed as the
last parameter of function arguments.